pyproject.toml
[project]
name = "routing_app"
version = "0.1.0"
description = "None"
readme = "README.md"
authors = [{ name = "pro2015", email = "" }]
requires-python = ">=3.12"
dependencies = [
    "fletxr==0.1.4.dev1",
    "flet[all]",
]